Cleanup storage and timing code Seperate storage and timing code into code and headers
